requirements.txt
flask>=2.0.1
flask-cors==3.0.10
google-generativeai==0.3.0
vectorstore-manager
langchain>=0.1.0
langchain-chroma==0.0.5  
langchain-community>=0.0.13
langchain-huggingface 
chromadb>=0.4.22
sentence-transformers>=2.2.2 
